+++ StereoSplitMode.cs
@@ -1,0 +1,7 @@
+namespace SpectrumViewerRT;
+
+public enum StereoSplitMode
+{
+    LeftRight,
+    TopBottom
+}